+++ expyriment/misc/_miscellaneous.py
@@ -302,32 +302,31 @@ def find_font(font):
         return ""
 
 
-def get_display_resolution(display=0):
-    """Return the display resolution.
+def get_display_info():
+    """Return information about the available display(s).
 
-    Parameters
-    ----------
-    display : int, optional
-        the display to get the resolution from (default=0)
-
-    Returns
-    -------
-    resolution : (int, int)
-        the display resolution
-
+    Return
+    ------
+    info : list of dicts
+        the information for each available display in the form
+            {"number": ...,              # display number, int
+             "maximal_resolution": ...,  # highest possible or "native", tuple
+             "desktop_resolution": ...}  # might differ (e.g. scaled), tuple
     """
 
-    from .. import _internals
-    if _internals.active_exp.is_initialized:
-        return _internals.active_exp.screen.display_resolution
-    else:
-        pygame.display.init()
-        return pygame.display.list_modes(display=display)[0]
+    info = []
+    for x in range(pygame.display.get_num_displays()):
+        info.append(
+            {"number": x,
+             "maximal_resolution": pygame.display.list_modes(display=x)[0],
+             "desktop_resolution": pygame.display.get_desktop_sizes()[x]})
+    return info
+
 
 def get_monitor_resolution():
     """Returns the monitor resolution
 
-    DEPRECATED! Use get_display_resolution instead.
+    DEPRECATED! Use get_display_info instead.
 
     Returns
     -------
@@ -335,8 +334,10 @@ def get_monitor_resolution():
         monitor resolution, screen resolution
 
     """
-
-    return get_display_resolution(display=0)
+    if _internals.active_exp.is_initialized:
+        return _internals.active_exp.screen.display_resolution
+    else:
+        return get_display_info[0]["maximal_resolution"]
 
 
 def is_ipython_running():